Edit existing files
This commit focus on the existing files and how they are modified to to
enable using the 'imx6q-ccimx6-bellequipment-mm3-development' MM3
development board.

The commit also include the modifications required to enable the 100mb
Ethernet driver and the correction to the physical device id. THIS might
be unneccessary, and SHOULD be changed later on to make the branch
compatible with linux main stream.

The IO controller device is also listed in the device tree compatibility
table.