Modifications to the device tree files
Edited the copied ccimx6sbc device tree files to work with the Bell
Equipment MM3 Development board. The naming of these files SHOULD change
in the future, dropping the '-development' to accomodate the production
versions, if we decide to stay with the i.MX6 platform.

These changes form the base device tree. The files will be duplicated
accross to u-boot also.